add exception and edit service / entity

+++ src/main/java/com/store/onlinestore/model/service/InventoryService.java
@@ -1,5 +1,6 @@
 package com.store.onlinestore.model.service;
 
+import com.store.onlinestore.controller.exception.InventoryNotFoundException;
 import com.store.onlinestore.model.entity.Inventory;
 import com.store.onlinestore.model.entity.Person;
 import com.store.onlinestore.model.entity.Product;
@@ -29,9 +30,13 @@ public class InventoryService {
     }
     public Inventory remove(Long id) throws Exception {
         try (CrudRepository<Inventory, Long> repository = new CrudRepository<>()) {
-            return repository.remove(id, Inventory.class);
+            if (repository.findById(id , Inventory.class) != null){
+                return repository.remove(id, Inventory.class);
+            }
+                throw new InventoryNotFoundException();
         }
     }
+
     public List<Inventory> findAll() throws Exception {
         try (CrudRepository<Inventory, Long> repository = new CrudRepository<>()) {
             return repository.findAll(Inventory.class);
